+++ fairlearn/preprocessing/__init__.py
@@ -1,3 +1,5 @@
-from ._correlation_remover import LinearDependenceRemover
+"""Preprocessing tools to help deal with sensitive attributes."""
+
+from ._linear_dep_remover import LinearDependenceRemover
 
 __all__ = ["LinearDependenceRemover"]
